add encryption package
