+++ encryption/aes/aes_test.go
@@ -1,11 +1,11 @@
-package common_test
+package aes_test
 
 import (
 	"encoding/hex"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/wego/pkg/common"
+	"github.com/wego/pkg/encryption/aes"
 )
 
 func Test_Encrypt_Decrypt_Ok(t *testing.T) {
@@ -14,10 +14,10 @@ func Test_Encrypt_Decrypt_Ok(t *testing.T) {
 	key := "1234567890qwertyuiop0123456789as"
 	plaintext := "Wego is awesome!"
 
-	ciphertext, err := common.Encrypt(plaintext, key)
+	ciphertext, err := aes.Encrypt(plaintext, key)
 	assert.NoError(err)
 
-	decrypted, err := common.Decrypt(ciphertext, key)
+	decrypted, err := aes.Decrypt(ciphertext, key)
 	assert.NoError(err)
 
 	assert.Equal(plaintext, decrypted)
@@ -29,7 +29,7 @@ func Test_Encrypt_KeyIsTooShort(t *testing.T) {
 	key := "1234567890qwertyuiop0123456789a"
 	plaintext := "Wego is awesome!"
 
-	ciphertext, err := common.Encrypt(plaintext, key)
+	ciphertext, err := aes.Encrypt(plaintext, key)
 	assert.Error(err)
 	assert.Empty(ciphertext)
 }
@@ -40,7 +40,7 @@ func Test_Decrypt_MalformedCiphertext_InvalidHexCode(t *testing.T) {
 	key := "1234567890qwertyuiop0123456789as"
 	ciphertext := "Wego is awesome!"
 
-	plaintext, err := common.Decrypt(ciphertext, key)
+	plaintext, err := aes.Decrypt(ciphertext, key)
 	assert.Error(err)
 	assert.Empty(plaintext)
 }
@@ -51,7 +51,7 @@ func Test_Decrypt_KeyIsTooShort(t *testing.T) {
 	key := "1234567890qwertyuiop0123456789a"
 	ciphertext := hex.EncodeToString([]byte{1, 2, 3, 4, 5, 6, 7, 8, 90})
 
-	plaintext, err := common.Decrypt(ciphertext, key)
+	plaintext, err := aes.Decrypt(ciphertext, key)
 	assert.Error(err)
 	assert.Empty(plaintext)
 }
@@ -62,7 +62,7 @@ func Test_Decrypt_MalformedCiphertext_InvalidNonceSize(t *testing.T) {
 	key := "1234567890qwertyuiop0123456789as"
 	ciphertext := hex.EncodeToString([]byte{1, 2, 3, 4, 5, 6, 7, 8, 90})
 
-	plaintext, err := common.Decrypt(ciphertext, key)
+	plaintext, err := aes.Decrypt(ciphertext, key)
 	assert.Error(err)
 	assert.Empty(plaintext)
 }
